+++ yael/vector.c
@@ -2302,5 +2302,5 @@ void b2fvec_cpy_subvectors (const unsigned char * v, int * idx, int d, int nout,
   long i, j;
   for (i = 0 ; i < nout ; i++)
     for (j = 0 ; j < d ; j++)
-      vout[i*d+j] = v[idx[i]*d+j];
+      vout[i*d+j] = v[idx[i] * (long) d+j];
 }
